+++ src/actions/fetch.js
@@ -1,7 +1,8 @@
 import { writeFile } from 'node:fs/promises';
+import { existsSync } from 'node:fs';
 import { DefaultAzureCredential } from '@azure/identity';
 import { ContainerClient } from '@azure/storage-blob';
-import { argParse, makeBlobStorageUrl, tryParseJSON } from '../common.js';
+import { argParse, log, makeBlobStorageUrl, tryParseJSON } from '../common.js';
 
 /** @import {ParseArgsConfig} from 'node:util' */
 
@@ -11,6 +12,7 @@ import { argParse, makeBlobStorageUrl, tryParseJSON } from '../common.js';
  * @prop {string} container
  * @prop {string} name
  * @prop {string} [output]
+ * @prop {boolean} [force]
  */
 
 
@@ -29,6 +31,10 @@ const OPTIONS = {
   name: {
     type: 'string',
     short: 'n',
+  },
+  force: {
+    type: 'boolean',
+    short: 'f',
   },
   output: {
     type: 'string',
@@ -54,6 +60,11 @@ export async function fetchHandler(args) {
   const contents = tryParseJSON(raw);
 
   if (config.output) {
+    if (existsSync(config.output) && !config.force) {
+      log(`file '${config.output}' already exitss. Use --force to overwrite`);
+      process.exit(1);
+    }
+
     await writeFile(config.output, contents);
   } else {
     console.log(contents);
